stamp/stamp1/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-serverside-datatable==2.0.0
mysqlclient==2.1.1
numpy==1.23.4
opencv-contrib-python==4.5.5.62
Pillow==9.2.0
sqlparse==0.4.3
tzdata==2022.7
